Name the root file of every crate after the crate to ease navigation
