+++ crates/chat_panel/Cargo.toml
@@ -3,6 +3,9 @@ name = "chat_panel"
 version = "0.1.0"
 edition = "2018"
 
+[lib]
+path = "src/chat_panel.rs"
+
 [dependencies]
 client = { path = "../client" }
 editor = { path = "../editor" }
