+++ crates/collections/Cargo.toml
@@ -3,6 +3,9 @@ name = "collections"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+path = "src/collections.rs"
+
 [features]
 test-support = ["seahash"]
