+++ crates/contacts_panel/src/contacts_panel.rs
