+++ crates/fsevent/Cargo.toml
@@ -4,6 +4,9 @@ version = "2.0.2"
 license = "MIT"
 edition = "2018"
 
+[lib]
+path = "src/fsevent.rs"
+
 [dependencies]
 bitflags = "1"
 fsevent-sys = "3.0.2"
