+++ crates/go_to_line/Cargo.toml
@@ -3,6 +3,9 @@ name = "go_to_line"
 version = "0.1.0"
 edition = "2018"
 
+[lib]
+path = "src/go_to_line.rs"
+
 [dependencies]
 text = { path = "../text" }
 editor = { path = "../editor" }
