+++ crates/language/Cargo.toml
@@ -3,6 +3,9 @@ name = "language"
 version = "0.1.0"
 edition = "2018"
 
+[lib]
+path = "src/language.rs"
+
 [features]
 test-support = [
     "rand",
