+++ crates/language/src/buffer.rs
@@ -1,15 +1,7 @@
-mod highlight_map;
-mod language;
-pub mod proto;
-#[cfg(test)]
-mod tests;
-
-pub use self::{
+pub use crate::{
     highlight_map::{HighlightId, HighlightMap},
-    language::{
-        BracketPair, Grammar, Language, LanguageConfig, LanguageRegistry, LanguageServerConfig,
-        PLAIN_TEXT,
-    },
+    proto, BracketPair, Grammar, Language, LanguageConfig, LanguageRegistry, LanguageServerConfig,
+    PLAIN_TEXT,
 };
 use anyhow::{anyhow, Result};
 use clock::ReplicaId;
@@ -73,7 +65,7 @@ pub struct Buffer {
     diagnostics_update_count: usize,
     language_server: Option<LanguageServerState>,
     #[cfg(test)]
-    operations: Vec<Operation>,
+    pub(crate) operations: Vec<Operation>,
 }
 
 pub struct Snapshot {
@@ -217,7 +209,7 @@ pub struct Chunk<'a> {
     pub diagnostic: Option<DiagnosticSeverity>,
 }
 
-struct Diff {
+pub(crate) struct Diff {
     base_version: clock::Global,
     new_text: Arc<str>,
     changes: Vec<(ChangeTag, usize)>,
@@ -573,7 +565,7 @@ impl Buffer {
         self.parse_count
     }
 
-    fn syntax_tree(&self) -> Option<Tree> {
+    pub(crate) fn syntax_tree(&self) -> Option<Tree> {
         if let Some(syntax_tree) = self.syntax_tree.lock().as_mut() {
             self.interpolate_tree(syntax_tree);
             Some(syntax_tree.tree.clone())
@@ -1094,7 +1086,7 @@ impl Buffer {
             .min_by_key(|(open_range, close_range)| close_range.end - open_range.start)
     }
 
-    fn diff(&self, new_text: Arc<str>, cx: &AppContext) -> Task<Diff> {
+    pub(crate) fn diff(&self, new_text: Arc<str>, cx: &AppContext) -> Task<Diff> {
         // TODO: it would be nice to not allocate here.
         let old_text = self.text();
         let base_version = self.version();
@@ -1111,7 +1103,7 @@ impl Buffer {
         })
     }
 
-    fn apply_diff(&mut self, diff: Diff, cx: &mut ModelContext<Self>) -> bool {
+    pub(crate) fn apply_diff(&mut self, diff: Diff, cx: &mut ModelContext<Self>) -> bool {
         if self.version == diff.base_version {
             self.start_transaction(None).unwrap();
             let mut offset = 0;
@@ -1153,7 +1145,7 @@ impl Buffer {
         self.start_transaction_at(selection_set_ids, Instant::now())
     }
 
-    fn start_transaction_at(
+    pub(crate) fn start_transaction_at(
         &mut self,
         selection_set_ids: impl IntoIterator<Item = SelectionSetId>,
         now: Instant,
@@ -1169,7 +1161,7 @@ impl Buffer {
         self.end_transaction_at(selection_set_ids, Instant::now(), cx)
     }
 
-    fn end_transaction_at(
+    pub(crate) fn end_transaction_at(
         &mut self,
         selection_set_ids: impl IntoIterator<Item = SelectionSetId>,
         now: Instant,
@@ -1995,7 +1987,7 @@ fn diagnostic_ranges<'a>(
         ))
 }
 
-fn contiguous_ranges(
+pub fn contiguous_ranges(
     values: impl IntoIterator<Item = u32>,
     max_len: usize,
 ) -> impl Iterator<Item = Range<u32>> {
