+++ crates/project_panel/src/project_panel.rs
