+++ crates/zed/Cargo.toml
@@ -7,7 +7,7 @@ version = "0.9.0"
 
 [lib]
 name = "zed"
-path = "src/lib.rs"
+path = "src/zed.rs"
 
 [[bin]]
 name = "Zed"
